refactor: remove legacy MathML parser setters and unused utility code
