+++ Gemfile
@@ -6,10 +6,10 @@ gemspec
 gem "mml"
 gem "rake", "~> 12.0"
 gem "rspec", "~> 3.0"
-gem "unitsml", "~> 0.6.2"
+gem "unitsml", github: "unitsml/unitsml-ruby", branch: "feat/context-register-models"
 gem 'canon'
 gem 'htmlentities'
-gem 'lutaml-model', '~> 0.8.0', github: "lutaml/lutaml-model", branch: "main"
+gem 'lutaml-model', '~> 0.8.0', github: "lutaml/lutaml-model", branch: "fix/global-context-register-lookup-fallback"
 gem 'oga'
 gem 'opal-rspec', "~> 1.1.0a"
 gem "rubocop-performance"
