+++ lib/plurimath/math/formula.rb
@@ -306,10 +306,6 @@ module Plurimath
         }
       end
 
-      def element_order=(value)
-        @value = validated_order(value.map(&:name))
-      end
-
       # Attributes start
       def mathcolor=(value)
         return if value.nil? || value.empty?
@@ -338,10 +334,6 @@ module Plurimath
         )
       end
 
-      def arg=(_); end
-
-      def scriptlevel=(_); end
-
       def intent=(value)
         return unless value
 
@@ -357,270 +349,12 @@ module Plurimath
       end
       # Attributes end
 
-      def ms_value=(value)
-        return if value.nil? || value.empty?
-
-        update(
-          replace_order_with_value(
-            @value,
-            Array(validate_symbols(value)),
-            "ms"
-          )
-        )
-      end
-
-      def mi_value=(value)
-        return if value.nil? || value.empty?
-
-        value = update_temp_mathml_values(value) if value.any? do |val|
-          val.is_a?(Math::Core) && val.temp_mathml_order.any?
-        end
-        update(
-          replace_order_with_value(
-            @value,
-            Array(validate_symbols(value)).flatten,
-            "mi"
-          )
-        )
-      end
-
-      def mn_value=(value)
-        return if value.nil? || value.empty?
-
-        update(
-          replace_order_with_value(
-            @value,
-            Array(validate_symbols(value)),
-            "mn"
-          )
-        )
-      end
-
-      def mtext_value=(value)
-        return if value.nil? || value.empty?
-
-        update(
-          replace_order_with_value(
-            @value,
-            Array(validate_symbols(value)),
-            "mtext"
-          )
-        )
-      end
-
-      def mo_value=(value)
-        return if value.nil? || value.empty?
-
-        value = update_temp_mathml_values(value)
-        update(
-          replace_order_with_value(
-            @value,
-            Array(value),
-            "mo"
-          )
-        )
-      end
-
-      def mstyle_value=(value)
-        return if value.nil? || value.empty?
-
-        replacing_order = value.length > 1 && value.any?(String)
-        update(
-          filter_values(
-            replace_order_with_value(
-              @value,
-              Array(
-                filter_values(
-                  value,
-                  array_to_instance: true,
-                  replacing_order: replacing_order,
-                ),
-              ),
-              "mstyle"
-            )
-          )
-        )
-      end
-
-      def mrow_value=(value)
-        return if value.nil? || value.empty?
-
-        replacing_order = value.length > 1 && value.any?(String)
-        update(
-          replace_order_with_value(
-            @value,
-            filter_values(
-              value,
-              array_to_instance: true,
-              replacing_order: replacing_order
-            ),
-            "mrow"
-          )
-        )
-      end
-
-      def munderover_value=(value)
-        update_temp_order(value, "munderover")
-      end
-
-      def msub_value=(value)
-        update_temp_order(value, "msub")
-      end
-
-      def msup_value=(value)
-        update_temp_order(value, "msup")
-      end
-
-      def mover_value=(value)
-        update_temp_order(value, "mover")
-      end
-
-      def munder_value=(value)
-        update_temp_order(value, "munder")
-      end
-
-      def msubsup_value=(value)
-        update_temp_order(value, "msubsup")
-      end
-
-      def mfrac_value=(value)
-        update_temp_order(value, "mfrac")
-      end
-
-      def msqrt_value=(value)
-        update_temp_order(value, "msqrt")
-      end
-
-      def mfenced_value=(value)
-        update_temp_order(value, "mfenced")
-      end
-
-      def mroot_value=(value)
-        update_temp_order(value, "mroot")
-      end
-
-      def msgroup_value=(value)
-        update_temp_order(value, "msgroup")
-      end
-
-      def mscarries_value=(value)
-        update_temp_order(value, "mscarries")
-      end
-
-      def msline_value=(value)
-        update_temp_order(value, "msline")
-      end
-
-      def msrow_value=(value)
-        update_temp_order(value, "msrow")
-      end
-
-      def semantics_value=(value)
-        update_temp_order(value, "semantics")
-      end
-
-      def mstack_value=(value)
-        update_temp_order(value, "mstack")
-      end
-
-      def merror_value=(value)
-        return if value.nil? || value.empty?
-
-        update(
-          replace_order_with_value(
-            @value,
-            filter_values(update_temp_mathml_values(value)),
-            "merror"
-          )
-        )
-      end
-
-      def mlongdiv_value=(value)
-        update_temp_order(value, "mlongdiv")
-      end
-
-      def none_value=(_)
-        @value&.delete("none")
-      end
-
-      def maligngroup_value=(value)
-        @value&.delete("maligngroup")
-      end
-
-      def menclose_value=(value)
-        update_temp_order(value, "menclose")
-      end
-
-      def mspace_value=(value)
-        return if value.nil? || value.empty?
-
-        if value.first.linebreak
-          linebreak = Math::Function::Linebreak.new(
-            nil,
-            { linebreak: value.first.linebreak }
-          )
-          update(
-            replace_order_with_value(
-              @value,
-              [linebreak],
-              "mspace"
-            )
-          )
-        else
-          @value&.delete("mspace")
-        end
-      end
-
-      def malignmark_value=(value)
-        @value&.delete("malignmark")
-      end
-
-      def mpadded_value=(value)
-        update_temp_order(value, "mpadded")
-      end
-
-      def mfraction_value=(value)
-        update_temp_order(value, "mfraction")
-      end
-
-      def mmultiscripts_value=(value)
-        update_temp_order(value, "mmultiscripts")
-      end
-
-      def mphantom_value=(value)
-        update_temp_order(value, "mphantom")
-      end
-
-      def mglyph_value=(value)
-        update_temp_order(value, "mglyph")
-      end
-
-      def a_value=(value)
-        @value&.delete("a")
-      end
-
       protected
 
       def asciimath_table_options(options, object)
         return options unless options.key?(:table) && object.is_a?(Math::Symbols::Comma)
 
         options.merge(literal_comma: true)
-      end
-
-      def update_temp_order(value, order_name)
-        return if value.nil? || value.empty?
-
-        update(
-          replace_order_with_value(
-            Array(@value),
-            Array(update_temp_mathml_values(value)),
-            order_name
-          )
-        )
-      end
-
-      def remove_order(order)
-        value.delete_if { |val| val.is_a?(String) && val == order }
       end
 
       def boolean_display_style(display_style = displaystyle)
