+++ lib/plurimath/math/function/longdiv.rb
@@ -5,7 +5,6 @@ module Plurimath
   module Math
     module Function
       class Longdiv < UnaryFunction
-        include Mathml::Utility
 
         def to_asciimath(options:)
           asciimath_value(options: options)
