+++ lib/plurimath/math/function/mlabeledtr.rb
@@ -5,8 +5,6 @@ module Plurimath
   module Math
     module Function
       class Mlabeledtr < BinaryFunction
-        include Mathml::Utility
-
         def to_mathml_without_math_tag(intent, options:)
           table = ox_element("mtable")
           mlabeledtr = ox_element(class_name)
@@ -17,22 +15,6 @@ module Plurimath
 
         def to_unicodemath(options:)
           "#{parameter_one&.to_unicodemath(options: options)}##{parameter_two&.value}"
-        end
-
-        def id=(value)
-          return if value.nil?
-
-          @parameter_two = Text.new(value)
-        end
-
-        def mtd_value=(value)
-          return if value.nil? || value.empty?
-
-          self.parameter_one = replace_order_with_value(
-            clear_temp_order,
-            update_temp_mathml_values(value),
-            "mtd"
-          )
         end
 
         protected
