+++ lib/plurimath/math/function/mpadded.rb
@@ -5,7 +5,6 @@ module Plurimath
   module Math
     module Function
       class Mpadded < UnaryFunction
-        include Mathml::Utility
 
         attr_accessor :options
