+++ lib/plurimath/math/function/multiscript.rb
@@ -5,8 +5,6 @@ module Plurimath
   module Math
     module Function
       class Multiscript < TernaryFunction
-        include Mathml::Utility
-
         FUNCTION = {
           name: "multiscript",
           first_value: "base",
@@ -79,16 +77,6 @@ module Plurimath
             )
             self.parameter_three = nil
           end
-        end
-
-        def none_value=(value)
-          return if value.nil? || value.empty?
-
-          @temp_mathml_order = replace_order_with_value(
-            @temp_mathml_order,
-            value,
-            "none"
-          )
         end
 
         private
