+++ lib/plurimath/math/function/none.rb
@@ -5,7 +5,6 @@ module Plurimath
   module Math
     module Function
       class None < UnaryFunction
-        include Mathml::Utility
 
         def to_asciimath(**); ""; end
