+++ lib/plurimath/math/function/semantics.rb
@@ -5,8 +5,6 @@ module Plurimath
   module Math
     module Function
       class Semantics < BinaryFunction
-        include Mathml::Utility
-
         FUNCTION = {
           name: "semantics",
           first_value: "first argument",
@@ -40,17 +38,6 @@ module Plurimath
             self.parameter_two = nil
             obj.update(semantics)
           end
-        end
-
-        def annotation=(value)
-          return unless value
-
-          @parameter_two = [
-            {
-              annotation: value,
-            }
-          ]
-          @temp_mathml_order.delete("annotation")
         end
 
         protected
