+++ lib/plurimath/math/function/td.rb
@@ -5,8 +5,6 @@ module Plurimath
   module Math
     module Function
       class Td < BinaryFunction
-        include Mathml::Utility
-
         def initialize(parameter_one = nil, parameter_two = nil)
           parameter_one&.delete_if { |td| td == "&" }
           super(Array(parameter_one), parameter_two)
