+++ lib/plurimath/math/function/tr.rb
@@ -5,8 +5,6 @@ module Plurimath
   module Math
     module Function
       class Tr < UnaryFunction
-        include Mathml::Utility
-
         def initialize(parameter_one = [])
           parameter_one.map!.with_index { |_, index| Td.new([]) } if parameter_one&.all?("@")
           super(parameter_one)
@@ -95,15 +93,6 @@ module Plurimath
           first_value
         end
 
-        def mtd_value=(value)
-          return if value.nil? || value.empty?
-
-          self.parameter_one = replace_order_with_value(
-            clear_temp_order,
-            update_temp_mathml_values(value),
-            "mtd"
-          )
-        end
       end
     end
   end
