+++ lib/plurimath/math/number.rb
@@ -4,7 +4,6 @@ module Plurimath
   module Math
     class Number < Core
       attr_accessor :value, :mini_sub_sized, :mini_sup_sized
-      include Mathml::Utility
 
       def initialize(value = nil, mini_sub_sized: false, mini_sup_sized: false)
         @value = value.is_a?(::Parslet::Slice) ? value.to_s : value
