--- lib/plurimath/mathml/utility/empty_defined_methods.rb
@@ -1,483 +1,0 @@
-# frozen_string_literal: true
-
-module Plurimath
-  class Mathml
-    module Utility
-      module EmptyDefinedMethods
-        def mi_value; end
-
-        def mo_value; end
-
-        def mn_value; end
-
-        def ms_value; end
-
-        def mtd_value; end
-
-        def mtr_value; end
-
-        def msub_value; end
-
-        def none_value; end
-
-        def msup_value; end
-
-        def mrow_value; end
-
-        def mroot_value; end
-
-        def mover_value; end
-
-        def mtext_value; end
-
-        def mfrac_value; end
-
-        def msrow_value; end
-
-        def msqrt_value; end
-
-        def mstack_value; end
-
-        def mspace_value; end
-
-        def mtable_value; end
-
-        def msline_value; end
-
-        def mstyle_value; end
-
-        def munder_value; end
-
-        def merror_value; end
-
-        def msubsup_value; end
-
-        def msgroup_value; end
-
-        def mfenced_value; end
-
-        def mpadded_value; end
-
-        def mphantom_value; end
-
-        def mlongdiv_value; end
-
-        def menclose_value; end
-
-        def mscarries_value; end
-
-        def semantics_value; end
-
-        def mlabeledtr_value; end
-
-        def munderover_value; end
-
-        def malignmark_value; end
-
-        def maligngroup_value; end
-
-        def mmultiscripts_value; end
-
-        def id; end
-
-        def dir; end
-
-        def color; end
-
-        def mathsize; end
-
-        def fontsize; end
-
-        def mathcolor; end
-
-        def fontstyle; end
-
-        def fontweight; end
-
-        def malignmark; end
-
-        def fontfamily; end
-
-        def background; end
-
-        def mathvariant; end
-
-        def scriptlevel; end
-
-        def decimalpoint; end
-
-        def displaystyle; end
-
-        def scriptminsize; end
-
-        def mathbackground; end
-
-        def mathbackgroundcolor; end
-
-        def infixlinebreakstyle; end
-
-        def scriptsizemultiplier; end
-
-        def accent; end
-
-        def annotation; end
-
-        def accentunder; end
-
-        def align; end
-
-        def alignmentscope; end
-
-        def bevelled; end
-
-        def charalign; end
-
-        def charspacing; end
-
-        def close; end
-
-        def columnalign; end
-
-        def columnlines; end
-
-        def columnspacing; end
-
-        def columnspan; end
-
-        def columnwidth; end
-
-        def crossout; end
-
-        def denomalign; end
-
-        def depth; end
-
-        def edge; end
-
-        def equalcolumns; end
-
-        def equalrows; end
-
-        def fence; end
-
-        def form; end
-
-        def frame; end
-
-        def framespacing; end
-
-        def groupalign; end
-
-        def height; end
-
-        def indentalign; end
-
-        def indentalignfirst; end
-
-        def indentalignlast; end
-
-        def indentshift; end
-
-        def indentshiftfirst; end
-
-        def indentshiftlast; end
-
-        def indenttarget; end
-
-        def largeop; end
-
-        def intent; end
-
-        def leftoverhang; end
-
-        def length; end
-
-        def linebreak; end
-
-        def linebreakmultchar; end
-
-        def linebreakstyle; end
-
-        def lineleading; end
-
-        def linethickness; end
-
-        def location; end
-
-        def longdivstyle; end
-
-        def lquote; end
-
-        def lspace; end
-
-        def maxsize; end
-
-        def minlabelspacing; end
-
-        def minsize; end
-
-        def movablelimits; end
-
-        def mslinethickness; end
-
-        def notation; end
-
-        def numalign; end
-
-        def open; end
-
-        def position; end
-
-        def rightoverhang; end
-
-        def rowalign; end
-
-        def rowlines; end
-
-        def rowspacing; end
-
-        def rowspan; end
-
-        def rquote; end
-
-        def rspace; end
-
-        def selection; end
-
-        def separator; end
-
-        def separators; end
-
-        def shift; end
-
-        def side; end
-
-        def stackalign; end
-
-        def stretchy; end
-
-        def subscriptshift; end
-
-        def superscriptshift; end
-
-        def symmetric; end
-
-        def valign; end
-
-        def width; end
-
-        def veryverythinmathspace; end
-
-        def verythinmathspace; end
-
-        def thinmathspace; end
-
-        def mediummathspace; end
-
-        def thickmathspace; end
-
-        def verythickmathspace; end
-
-        def veryverythickmathspace; end
-
-        def mscarry_value; end
-
-        def mfraction_value; end
-
-        def mprescripts_value; end
-
-        def mglyph_value; end
-
-        def __root; end
-
-        def __parent; end
-
-        def __root=(_); end
-
-        def __parent=(_); end
-
-        def malignmark=(_); end
-
-        def mathcolor=(_); end
-
-        def mathbackground=(_); end
-
-        def mathsize=(_); end
-
-        def dir=(_); end
-
-        def fontfamily=(_); end
-
-        def fontweight=(_); end
-
-        def fontstyle=(_); end
-
-        def fontsize=(_); end
-
-        def color=(_); end
-
-        def mathbackgroundcolor=(_); end
-
-        def background=(_); end
-
-        def scriptlevel=(_); end
-
-        def displaystyle=(_); end
-
-        def scriptsizemultiplier=(_); end
-
-        def scriptminsize=(_); end
-
-        def infixlinebreakstyle=(_); end
-
-        def decimalpoint=(_); end
-
-        def align=(_); end
-
-        def alignmentscope=(_); end
-
-        def charalign=(_); end
-
-        def charspacing=(_); end
-
-        def close=(_); end
-
-        def columnalign=(_); end
-
-        def columnlines=(_); end
-
-        def columnspacing=(_); end
-
-        def columnspan=(_); end
-
-        def columnwidth=(_); end
-
-        def crossout=(_); end
-
-        def denomalign=(_); end
-
-        def depth=(_); end
-
-        def edge=(_); end
-
-        def equalcolumns=(_); end
-
-        def equalrows=(_); end
-
-        def fence=(_); end
-
-        def form=(_); end
-
-        def frame=(_); end
-
-        def framespacing=(_); end
-
-        def groupalign=(_); end
-
-        def height=(_); end
-
-        def indentalign=(_); end
-
-        def indentalignfirst=(_); end
-
-        def indentalignlast=(_); end
-
-        def indentshift=(_); end
-
-        def indentshiftfirst=(_); end
-
-        def indentshiftlast=(_); end
-
-        def indenttarget=(_); end
-
-        def intent=(_); end
-
-        def largeop=(_); end
-
-        def leftoverhang=(_); end
-
-        def length=(_); end
-
-        def linebreakmultchar=(_); end
-
-        def lineleading=(_); end
-
-        def location=(_); end
-
-        def longdivstyle=(_); end
-
-        def lquote=(_); end
-
-        def lspace=(_); end
-
-        def maxsize=(_); end
-
-        def minlabelspacing=(_); end
-
-        def minsize=(_); end
-
-        def movablelimits=(_); end
-
-        def mslinethickness=(_); end
-
-        def numalign=(_); end
-
-        def open=(_); end
-
-        def position=(_); end
-
-        def rightoverhang=(_); end
-
-        def rowalign=(_); end
-
-        def rowlines=(_); end
-
-        def rowspacing=(_); end
-
-        def rowspan=(_); end
-
-        def rquote=(_); end
-
-        def selection=(_); end
-
-        def separator=(_); end
-
-        def separators=(_); end
-
-        def shift=(_); end
-
-        def side=(_); end
-
-        def stackalign=(_); end
-
-        def stretchy=(_); end
-
-        def subscriptshift=(_); end
-
-        def superscriptshift=(_); end
-
-        def symmetric=(_); end
-
-        def valign=(_); end
-
-        def width=(_); end
-
-        def veryverythinmathspace=(_); end
-
-        def verythinmathspace=(_); end
-
-        def thinmathspace=(_); end
-
-        def mediummathspace=(_); end
-
-        def thickmathspace=(_); end
-
-        def verythickmathspace=(_); end
-
-        def veryverythickmathspace=(_); end
-
-        def mprescripts_value=(_); end
-      end
-    end
-  end
-end
